+++ src/components/admin/materii/constants.js
@@ -1,0 +1,29 @@
+export const facultati = [
+  "Facultatea de Matematică și Informatică",
+  "Facultatea de Fizică",
+  // ... alte facultăți
+];
+
+export const specializari = {
+  "Facultatea de Matematică și Informatică": ["IR", "IG", "MI", "MA"],
+  // ... alte specializări pentru alte facultăți
+};
+
+export const ani = ["I", "II", "III"];
+export const semestre = ["1", "2"];
+
+// Funny Romanian loading messages
+export const funnyLoadingMessages = [
+  "🤖 Învăț să citesc PDF-uri ca un profesor în primul an...",
+  "📚 Caut materiile prin labirintul academic...",
+  "🔍 Analizez planul de învățământ cu lupa...",
+  "💭 Mă gândesc mai tare decât la examenul de licență...",
+  "🎯 Extrag datele cu precizia unui chirurg...",
+  "🚀 Zbor prin semestre ca Superman prin nori...",
+  "🧠 Fac gimnastică mentală cu credite și semestre...",
+  "📊 Organizez materiile ca un biblioteccar obsesiv...",
+  "⚡ Procesez informațiile cu viteza internetului din cămin...",
+  "🎪 Jonglез cu materii obligatorii și opționale...",
+  "🔥 Ard etapele ca un student în sesiune...",
+  "🎨 Pictez un tabel perfect cu materiile tale...",
+];
